requirements.txt
absl-py==1.4.0
altair==5.1.1
astunparse==1.6.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
flatbuffers==23.5.26
fonttools==4.42.1
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.35
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.0
Pillow==9.5.0
protobuf==4.24.3
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rpds-py==0.10.2
rsa==4.9
six==1.16.0
smmap==5.0.0
streamlit==1.26.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.22.0
watchdog==3.0.0
Werkzeug==2.3.7
wrapt==1.15.0
zipp==3.16.2
